Refactor core module as common
