+++ components/org.wso2.carbon.identity.application.authenticator.push.common/src/main/java/org/wso2/carbon/identity/application/authenticator/push/common/exception/IdentityPushAuthException.java
@@ -17,16 +17,23 @@
  *
  */
 
-package org.wso2.carbon.identity.application.authenticator.push.core.exception;
+package org.wso2.carbon.identity.application.authenticator.push.common.exception;
 
-public class PushAuthTokenValidationException extends IdentityPushAuthException {
+import org.wso2.carbon.identity.base.IdentityException;
 
-    public PushAuthTokenValidationException(String message) {
+/**
+ * Exception class for Push authentication core module.
+ */
+public class IdentityPushAuthException extends IdentityException {
+
+    private static final long serialVersionUID = 1481059218426478592L;
+
+    public IdentityPushAuthException(String message) {
 
         super(message);
     }
 
-    public PushAuthTokenValidationException(String message, Throwable e) {
+    public IdentityPushAuthException(String message, Throwable e) {
 
         super(message, e);
     }
