+++ components/org.wso2.carbon.identity.application.authenticator.push.common/src/main/java/org/wso2/carbon/identity/application/authenticator/push/common/exception/PushAuthTokenValidationException.java
@@ -17,23 +17,16 @@
  *
  */
 
-package org.wso2.carbon.identity.application.authenticator.push.core.exception;
+package org.wso2.carbon.identity.application.authenticator.push.common.exception;
 
-import org.wso2.carbon.identity.base.IdentityException;
+public class PushAuthTokenValidationException extends IdentityPushAuthException {
 
-/**
- * Exception class for Push authentication core module.
- */
-public class IdentityPushAuthException extends IdentityException {
-
-    private static final long serialVersionUID = 1481059218426478592L;
-
-    public IdentityPushAuthException(String message) {
+    public PushAuthTokenValidationException(String message) {
 
         super(message);
     }
 
-    public IdentityPushAuthException(String message, Throwable e) {
+    public PushAuthTokenValidationException(String message, Throwable e) {
 
         super(message, e);
     }
